log: more robust log manipulation for shortcuts
The current log replacement for shortcuts does not work well when hashes are
referred to inside revision descriptions.

This change does 2 things:
    - Uses a more robust template string that does not exhibit the above bug
    - Drops one of the two VCS calls necessary to dramatically improve
      performance in big repositories

This requires the addition of a dependency on `regex`, which is one of the
best crates in the ecosystem and will most likely be used in other future
patches.